Party/User Relations

+++ backend/src/main/java/org/partypets/backend/service/PartyService.java
@@ -4,6 +4,9 @@ import org.partypets.backend.model.PartyWithoutId;
 import org.partypets.backend.model.Party;
 import org.partypets.backend.model.UuIdService;
 import org.partypets.backend.repo.PartyRepo;
+import org.partypets.backend.security.MongoUser;
+import org.partypets.backend.security.MongoUserService;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,18 +18,23 @@ public class PartyService {
 
     private final UuIdService uuIdService;
 
-    public PartyService(PartyRepo partyRepo, UuIdService uuIdService) {
+    private final MongoUserService userService;
+
+
+    public PartyService(PartyRepo partyRepo, UuIdService uuIdService, MongoUserService userService) {
         this.partyRepo = partyRepo;
         this.uuIdService = uuIdService;
+        this.userService = userService;
     }
 
     public List<Party> list() {
         return this.partyRepo.findAll();
     }
-
     public Party add(PartyWithoutId newParty) {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        MongoUser user = this.userService.getUserByUsername(username);
         String id = uuIdService.getRandomId();
-        Party party = new Party(id, newParty.getDate(), newParty.getLocation(), newParty.getTheme());
+        Party party = new Party(id, newParty.getDate(), newParty.getLocation(), newParty.getTheme(), user.id());
         return this.partyRepo.insert(party);
     }
 
@@ -34,13 +42,25 @@ public class PartyService {
         return this.partyRepo.findById(id).orElseThrow();
     }
 
-
     public Party edit(String id, PartyWithoutId newParty) {
-        Party editedParty = new Party(id, newParty.getDate(), newParty.getLocation(), newParty.getTheme());
-        return this.partyRepo.save(editedParty);
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        MongoUser user = this.userService.getUserByUsername(username);
+
+        Party currentParty = this.partyRepo.findById(id).orElseThrow();
+        if (currentParty.getUserId().equals(user.id())) {
+            Party editedParty = new Party(id, newParty.getDate(), newParty.getLocation(), newParty.getTheme(), user.id());
+            return this.partyRepo.save(editedParty);
+        }
+        return currentParty;
     }
 
     public void delete(String id) {
-        this.partyRepo.deleteById(id);
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        MongoUser user = this.userService.getUserByUsername(username);
+
+        Party currentParty = this.partyRepo.findById(id).orElseThrow();
+        if (currentParty.getUserId().equals(user.id())) {
+            this.partyRepo.deleteById(id);
+        }
     }
 }
